Protegendo rotas de criação de posts e salvando token jwt em cookie

+++ backend/src/main/java/com/rafael_souza_de_almeida/inTempo/Service/PostService.java
@@ -8,9 +8,11 @@ import com.rafael_souza_de_almeida.inTempo.Exception.PostNotFoundException;
 import com.rafael_souza_de_almeida.inTempo.Exception.UserNotFoundException;
 import com.rafael_souza_de_almeida.inTempo.Repository.PostRepository;
 import com.rafael_souza_de_almeida.inTempo.Repository.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 
 @Service
@@ -18,10 +20,12 @@ public class PostService {
 
     private final PostRepository postRepository;
     private final UserRepository userRepository;
+    private final JwtService jwtService;
 
-    public PostService(PostRepository postRepository, UserRepository userRepository) {
+    public PostService(PostRepository postRepository, UserRepository userRepository, JwtService jwtService) {
         this.postRepository = postRepository;
         this.userRepository = userRepository;
+        this.jwtService = jwtService;
     }
 
     public List<PostDTO> findAllPosts() {
@@ -29,9 +33,11 @@ public class PostService {
         return posts.stream().map(PostDTO::new).toList();
     }
 
-    public PostDTO save(CreateOrEditPostDTO dto) throws UserNotFoundException {
+    public PostDTO save(CreateOrEditPostDTO dto, HttpServletRequest request) throws UserNotFoundException {
 
-        User user = userRepository.findById(dto.getUser_id()).orElseThrow(() -> new UserNotFoundException("Usuário não encontrado"));
+        String user_id = jwtService.extractIdFromCookie(request);
+
+        User user = userRepository.findById(user_id).orElseThrow(() -> new UserNotFoundException("Usuário não encontrado"));
 
         Post post = new Post();
 
@@ -45,21 +51,28 @@ public class PostService {
 
     }
 
-    public void delete(Long id) throws PostNotFoundException {
+    public void delete(Long id, HttpServletRequest request) throws PostNotFoundException, AccessDeniedException {
         Post post = postRepository.findById(id).orElseThrow(() -> new PostNotFoundException("Post não encontrado"));
+
+        if(!post.getUser().getId().equals(jwtService.extractIdFromCookie(request))) {
+
+            throw new AccessDeniedException("Acesso negado");
+        }
 
         postRepository.delete(post);
 
     }
 
     @Transactional
-    public PostDTO update(CreateOrEditPostDTO dto, Long id) throws PostNotFoundException {
+    public PostDTO update(CreateOrEditPostDTO dto, Long id, HttpServletRequest request) throws PostNotFoundException, AccessDeniedException {
 
         Post postToUpdate = postRepository.findById(id).orElseThrow(()-> new PostNotFoundException("Post não encontrado"));
 
-        if(dto.getContent() != null) {
-            postToUpdate.setContent(dto.getContent());
+        if(!postToUpdate.getUser().getId().equals(jwtService.extractIdFromCookie(request))) {
+            throw new AccessDeniedException("Acesso negado");
         }
+
+        postToUpdate.setContent(dto.getContent());
 
         postRepository.save(postToUpdate);
